Implemented types for all the components
